+++ .github/workflows/ci.yml
@@ -25,8 +25,8 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
       - uses: Swatinem/rust-cache@v1
       - run: >
-          cargo run -p starknet_node --bin starknet_client_integration_test;
-          cargo run -p starknet_node --bin central_source_integration_test
+          cargo run -p papyrus_node --bin starknet_client_integration_test;
+          cargo run -p papyrus_node --bin central_source_integration_test
 
   rustfmt:
     runs-on: ubuntu-latest
@@ -38,6 +38,40 @@ jobs:
           toolchain: nightly-2022-07-27
       - uses: Swatinem/rust-cache@v1
       - run: cargo +nightly-2022-07-27 fmt --all -- --check
+
+  udeps:
+    runs-on: ubuntu-latest
+    steps:
+      # get the code
+      - uses: actions/checkout@v1
+        name: "Repository Checkout"
+      # set up rust environment
+      - uses: actions-rs/toolchain@v1
+        name: "Rust Toolchain Setup"
+        with:
+          profile: minimal
+          toolchain: nightly-2022-07-27
+          override: true
+      # cache build to wait shorter on next builds
+      - uses: actions/cache@v2
+        name: "Cache Setup"
+        with:
+          path: |
+            ~/.cargo/registry
+            ~/.cargo/git
+            target
+          key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
+      # install cargo-udeps
+      - uses: actions-rs/cargo@v1
+        name: "Udeps Installation"
+        with:
+          command: install
+          args: cargo-udeps --locked
+      # use cargo-udeps
+      - uses: actions-rs/cargo@v1
+        name: "Unused Dependency Check"
+        with:
+          command: udeps
 
   clippy:
     runs-on: ubuntu-latest
