+++ crates/papyrus_gateway/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "papyrus_gateway"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { version = "1.0.44" }
+bincode = { version = "1.3.3" }
+papyrus_storage = { path = "../papyrus_storage" }
+futures-util = { version = "0.3.21" }
+jsonrpsee = { version = "0.13.1", features = ["full"] }
+log = { version = "0.4" }
+serde = { version = "1.0.130", features = ["derive"] }
+serde_json = { version = "1.0.81" }
+starknet_api = { path = "../starknet_api" }
+thiserror = { version = "1.0.31" }
+tokio = { version = "1.18.2", features = ["full","sync"] }
+tokio-stream = { version = "0.1.8" }
+url = { version = "2.2.2" }
+
+[dev-dependencies]
+assert_matches = { version = "1.5.0" }
+tempfile = { version = "3.3.0" }
+papyrus_storage = { path = "../papyrus_storage", features = ["testing"] }
