+++ crates/papyrus_node/src/bin/starknet_client_integration_test.rs
@@ -1,7 +1,7 @@
+use papyrus_node::config::load_config;
 use starknet_api::serde_utils::bytes_from_hex_str;
 use starknet_api::{BlockNumber, ClassHash, StarkHash};
 use starknet_client::{StarknetClient, StarknetClientTrait};
-use starknet_node::config::load_config;
 
 #[tokio::main]
 async fn main() {
