+++ crates/papyrus_node/src/lib.rs
@@ -1,0 +1,2 @@
+#[allow(unused_imports)]
+pub mod config;
