+++ crates/papyrus_storage/Cargo.toml
@@ -1,21 +1,26 @@
 [package]
-name = "starknet_client"
+name = "papyrus_storage"
 version = "0.1.0"
 edition = "2021"
 
+[features]
+testing = []
+
 [dependencies]
-async-trait = { version = "0.1.56" }
+bincode = { version = "1.3.3" }
+futures-util = { version = "0.3.21" }
+libmdbx = { version = "0.1.5", features = ["lifetimed-bytes"] }
 log = { version = "0.4" }
 reqwest = { version = "0.11", features = ["json", "blocking"] }
 serde = { version = "1.0.130", features = ["derive"] }
-serde_json = { version = "1.0.81" , features = ["arbitrary_precision"]}
+serde_json = { version = "1.0.81" }
 starknet_api = { path = "../starknet_api" }
 thiserror = { version = "1.0.31" }
 tokio = { version = "1.18.2", features = ["full","sync"] }
+tokio-stream = { version = "0.1.8" }
 url = { version = "2.2.2" }
-web3 = { version = "0.18.0" }
+tempfile = { version = "3.3.0" }
 
 [dev-dependencies]
-assert = { version = "0.0.4" }
+anyhow = { version = "1.0.44" }
 assert_matches = { version = "1.5.0" }
-mockito = { version = "0.31.0" }
