+++ crates/papyrus_storage/src/test_utils.rs
@@ -1,0 +1,16 @@
+use tempfile::tempdir;
+
+use super::{open_storage, StorageReader, StorageWriter};
+use crate::db::DbConfig;
+
+pub fn get_test_config() -> DbConfig {
+    let dir = tempdir().unwrap();
+    DbConfig {
+        path: dir.path().to_str().unwrap().to_string(),
+        max_size: 1 << 35, // 32GB.
+    }
+}
+pub fn get_test_storage() -> (StorageReader, StorageWriter) {
+    let config = get_test_config();
+    open_storage(config).expect("Failed to open storage.")
+}
