--- crates/starknet_node/src/lib.rs
@@ -1,5 +1,0 @@
-#[allow(unused_imports)]
-pub mod config;
-pub mod gateway;
-pub mod storage;
-pub mod sync;
